requirements.txt
Django>=5.2.6
django-environ
django-cors-headers
django-storages
djangorestframework
django-phonenumber-field
phonenumbers
django-filter
django-debug-toolbar
django-rosetta
django-import-export
drf-spectacular
django-celery-beat
django-push-notifications
celery
redis
psycopg2-binary
boto3
firebase-admin
Pillow
python-decouple
num2words
python-magic
pandas
weasyprint
drf-excel
drf-nested-routers
